+++ sdk/daml-lf/validation/src/test/scala/com/digitalasset/daml/lf/validation/upgrade/UpgradesSpecAdminAPIWithValidation.scala
@@ -1,0 +1,7 @@
+package com.daml.lf.validation.upgrade
+
+
+class UpgradesSpecAdminAPIWithValidation
+  extends UpgradesSpecAdminAPI("Admin API with validation")
+    with ShortTests
+    with LongTests
