+++ sdk/daml-lf/validation/src/test/scala/com/digitalasset/daml/lf/validation/upgrade/UpgradesSpecBase.scala
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 package com.daml.lf.validation
+package upgrade
 
 import com.daml.SdkVersion
 import com.daml.bazeltools.BazelRunfiles
@@ -25,27 +26,6 @@ import scala.concurrent.Future
 import scala.io.Source
 import scala.util.{Success, Failure}
 
-class UpgradesSpecAdminAPIWithoutValidation
-    extends UpgradesSpecAdminAPI("Admin API without validation")
-    with ShortTests {
-  override val disableUpgradeValidation = true
-}
-
-class UpgradesSpecLedgerAPIWithoutValidation
-    extends UpgradesSpecLedgerAPI("Ledger API without validation")
-    with ShortTests {
-  override val disableUpgradeValidation = true
-}
-
-class UpgradesSpecAdminAPIWithValidation
-    extends UpgradesSpecAdminAPI("Admin API with validation")
-    with ShortTests
-    with LongTests
-
-class UpgradesSpecLedgerAPIWithValidation
-    extends UpgradesSpecLedgerAPI("Ledger API with validation")
-    with ShortTests
-    with LongTests
 
 abstract class UpgradesSpecAdminAPI(override val suffix: String) extends UpgradesSpec(suffix) {
   override def uploadPackage(
@@ -65,7 +45,7 @@ abstract class UpgradesSpecAdminAPI(override val suffix: String) extends Upgrade
   }
 }
 
-class UpgradesSpecLedgerAPI(override val suffix: String = "Ledger API")
+abstract class UpgradesSpecLedgerAPI(override val suffix: String = "Ledger API")
     extends UpgradesSpec(suffix) {
 
   override def uploadPackage(
