+++ sdk/daml-lf/validation/src/test/scala/com/digitalasset/daml/lf/validation/upgrade/UpgradesSpecLedgerAPIWithValidation.scala
@@ -1,0 +1,7 @@
+package com.daml.lf.validation.upgrade
+
+
+class UpgradesSpecLedgerAPIWithValidation
+  extends UpgradesSpecLedgerAPI("Ledger API with validation")
+    with ShortTests
+    with LongTests
